add whitelist to securityConfig
This splits the WebSecurityConfigTest into a test using the default value of the whitelist property.
This should ensure the known behaviour.
A second test uses a set whitelist for overwriting the actuator allowence.
This ensures the new behaviour.
Removed public keyword from touched test method/classes to improve readability as suggested in java:S5786.